Update README instruction for adding Shiny server
I've been trying to add the Shiny server to an RStudio container and struggled with the current instructions. I noticed some updated instructions from @cboettig in #219 given this was the error I was getting. 

It seems that

    RUN /rocker_scripts/install_shiny_server.sh

will work in all versions, so I've updated the instruction in the README to reflect this. Hope it might help someone else.